Move functions to skills.ts file
