+++ pages/api/skill.ts
@@ -41,6 +41,47 @@ export const getSkillsForTopic = (topic: Topic): Skill[] => {
   }
   return [];
 };
+export function getEmoji(emojiNum: number | null) {
+  if (emojiNum == null) {
+    return "❓";
+  } else if (emojiNum >= 0 && emojiNum <= 33) {
+    return "😔";
+  } else if (emojiNum >= 34 && emojiNum <= 66) {
+    return "😐";
+  } else {
+    return "😄";
+  }
+}
+export function getSkillId(skill: Skill) {
+  switch (skill) {
+    case Skill.ADDITION_SINGLE:
+      return 1;
+    case Skill.ADDITION_DOUBLE:
+      return 2;
+    case Skill.ADDITION_TRIPLE:
+      return 3;
+    case Skill.ADDITION_PROPERTIES:
+      return 4;
+    case Skill.SUBTRACTION_SINGLE:
+      return 34;
+    case Skill.SUBTRACTION_DOUBLE:
+      return 35;
+    case Skill.SUBTRACTION_TRIPLE:
+      return 36;
+    case Skill.EQUAL_GROUP_10_ITEMS:
+      return 37;
+    case Skill.MULTIPLICATION_5:
+      return 38;
+    case Skill.MULTIPLICATION_10:
+      return 39;
+    case Skill.EQUAL_SHARING_8_ITEMS:
+      return 40;
+    case Skill.DIVIDE_12_EQUALLY:
+      return 41;
+    case Skill.DIVIDE_100:
+      return 42;
+  }
+}
 
 export const getSkillsForTopicGrade = (topic: Topic, grade: Grade): Skill[] => {
   if (topic == Topic.NUMBERS) {
